phase 3 of issue #330
